Get rid of the search initializer and move to a simple predicate
